fix(filter): bound stripField recursion, prune empty composites
